star grafo

+++ Trabalho IA - Poupador/src/algoritmo/utils/aresta.java
@@ -2,11 +2,11 @@
 package algoritmo.utils;
 
 public class aresta {
-    public float peso;
+    public double peso;
     public int destino ;
     public int origem;
     public int dir;
-    public aresta(float peso,int destino,int origem, int dir){
+    public aresta(double peso,int destino,int origem, int dir){
         this.peso=peso;
         this.destino=destino;
         this.origem=origem;
